+++ include/cpfusa/fusa.hpp
@@ -9,10 +9,10 @@
 
 namespace cpfusa {
 
-constexpr std::string_view Version      = "0.12.6";
+constexpr std::string_view Version      = "0.13.0";
 constexpr std::string_view VersionMajor = "0";
-constexpr std::string_view VersionMinor = "12";
-constexpr std::string_view VersionPatch = "6";
+constexpr std::string_view VersionMinor = "13";
+constexpr std::string_view VersionPatch = "0";
 constexpr std::string_view SpecVersion  = "1.10.4";
 
 enum class Severity { INFO, WARNING, ERROR };